### 2023-06-11 #5538 @editorialbot
[PRE REVIEW]: SBArchOpt: Surrogate-Based Architecture Optimization
**Submitting author:** @jbussemaker (<a href="http://orcid.org/0000-0002-5421-6419">Jasper Bussemaker</a>)
**Repository:** https://github.com/jbussemaker/SBArchOpt
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v1.1.0
