+++ paper.bib
@@ -72,7 +72,6 @@ year = {2015}
   year      = {2022},
   address   = {Detroit, MI, USA},
   month     = jun,
-  doi       = {10.1002/iis2.12935},
 }
 
 @inproceedings{Balandat2020,
@@ -135,6 +134,7 @@ year = {2015}
   booktitle={American Institute of Aeronautics and Astronautics ({AIAA}) Aviation 2022 Forum},
   pages={3899},
   year={2022}
+  doi       = {10.2514/6.2022-3899},
 }
 
 @InProceedings{Apaza2021,
