+++ paper.bib
@@ -133,7 +133,7 @@ year = {2015}
   author={Bekemeyer, P. and Bertram, A. and Hines Chaves, D.A. and Dias Ribeiro, M. and Garbo, A. and Kiener, A. and Sabater, C. and Stradtner, M. and Wassing, S. and Widhalm, M. and others},
   booktitle={American Institute of Aeronautics and Astronautics ({AIAA}) Aviation 2022 Forum},
   pages={3899},
-  year={2022}
+  year={2022},
   doi       = {10.2514/6.2022-3899},
 }
 
