+++ README.md
@@ -5,6 +5,9 @@
 [![License](https://img.shields.io/badge/license-MIT-green.svg)](LICENSE)
 [![status](https://joss.theoj.org/papers/0b2b765c04d31a4cead77140f82ecba0/status.svg)](https://joss.theoj.org/papers/0b2b765c04d31a4cead77140f82ecba0)
 
+[GitHub Repository](https://github.com/jbussemaker/SBArchOpt) |
+[Documentation](https://github.com/jbussemaker/SBArchOpt/blob/main/docs/readme.md)
+
 SBArchOpt (es-bee-ARK-opt) provides a set of classes and interfaces for applying Surrogate-Based Optimization (SBO)
 for system architecture optimization problems:
 - Expensive black-box problems: evaluating one candidate architecture might be computationally expensive
@@ -46,19 +49,23 @@ Refer to their documentation for dedicated installation instructions.
 
 ## Documentation
 
-Refer to the [documentation](docs/readme.md) for more background on SBArchOpt
+Refer to the [documentation](https://github.com/jbussemaker/SBArchOpt/blob/main/docs/readme.md) for more background on SBArchOpt
 and how to implement architecture optimization problems.
-Test problem documentation can be found here: [test problems](docs/test_problems.md)
+Test problem documentation can be found here: [test problems](https://github.com/jbussemaker/SBArchOpt/blob/main/docs/test_problems.md)
 
 Optimization framework documentation:
-- [pymoo](docs/algo_pymoo.md)
-- [ArchSBO](docs/algo_arch_sbo.md)
-- [BoTorch (Ax)](docs/algo_botorch.md)
-- [Trieste](docs/algo_trieste.md)
-- [HEBO](docs/algo_hebo.md)
-- [TPE](docs/algo_tpe.md)
-- [SEGOMOE](docs/algo_segomoe.md)
-- [SMARTy](docs/algo_smarty.md)
+- [pymoo](https://github.com/jbussemaker/SBArchOpt/blob/main/docs/algo_pymoo.md)
+- [ArchSBO](https://github.com/jbussemaker/SBArchOpt/blob/main/docs/algo_arch_sbo.md)
+- [BoTorch (Ax)](https://github.com/jbussemaker/SBArchOpt/blob/main/docs/algo_botorch.md)
+- [Trieste](https://github.com/jbussemaker/SBArchOpt/blob/main/docs/algo_trieste.md)
+- [HEBO](https://github.com/jbussemaker/SBArchOpt/blob/main/docs/algo_hebo.md)
+- [TPE](https://github.com/jbussemaker/SBArchOpt/blob/main/docs/algo_tpe.md)
+- [SEGOMOE](https://github.com/jbussemaker/SBArchOpt/blob/main/docs/algo_segomoe.md)
+- [SMARTy](https://github.com/jbussemaker/SBArchOpt/blob/main/docs/algo_smarty.md)
+
+See also the tutorials:
+- [SBArchOpt Tutorial](https://github.com/jbussemaker/SBArchOpt/blob/main/docs/tutorial.ipynb): optimization, implementing new problems
+- [Tunable Hierarchical Meta Problem Tutorial](https://github.com/jbussemaker/SBArchOpt/blob/main/docs/tutorial_tunable_meta_problem.ipynb)
 
 ## Contributing
 
