### 2023-06-12 d5fc21b
Ensure not too many design vectors are generated
