+++ README.md
@@ -79,4 +79,6 @@ Contributions are appreciated too:
 - Add your contributions to the fork
   - Update/add documentation
   - Add tests and make sure they pass (tests are run using `pytest`)
+- Read and sign the [Contributor License Agreement (CLA)](https://github.com/jbussemaker/SBArchOpt/blob/main/SBArchOpt%20DLR%20Individual%20Contributor%20License%20Agreement.docx)
+  , and send it to the project coordinator
 - Issue a pull request
