### 2023-06-14 1aca214
Update constraint handling API; add constraint elimination
