+++ sb_arch_opt/algo/arch_sbo/infill.py
@@ -17,6 +17,7 @@ Contact: jasper.bussemaker@dlr.de
 import timeit
 import numpy as np
 from typing import *
+from enum import Enum
 from scipy.stats import norm
 from scipy.special import ndtr
 from sb_arch_opt.problem import ArchOptProblemBase
@@ -32,7 +33,8 @@ from pymoo.algorithms.moo.nsga2 import RankAndCrowdingSurvival, calc_crowding_di
 __all__ = ['SurrogateInfill', 'FunctionEstimateInfill', 'ConstrainedInfill', 'FunctionEstimateConstrainedInfill',
            'ExpectedImprovementInfill', 'MinVariancePFInfill', 'ConstraintStrategy', 'MeanConstraintPrediction',
            'ProbabilityOfFeasibility', 'ProbabilityOfImprovementInfill', 'LowerConfidenceBoundInfill',
-           'MinimumPoIInfill', 'EnsembleInfill', 'IgnoreConstraints', 'get_default_infill', 'HCInfill']
+           'MinimumPoIInfill', 'EnsembleInfill', 'IgnoreConstraints', 'get_default_infill', 'HCInfill',
+           'ConstraintAggregation']
 
 try:
     from smt.surrogate_models.surrogate_model import SurrogateModel
@@ -41,8 +43,8 @@ except ImportError:
     pass
 
 
-def get_default_infill(problem: ArchOptProblemBase, n_parallel: int = None, min_pof: float = None) \
-        -> Tuple['ConstrainedInfill', int, bool]:
+def get_default_infill(problem: ArchOptProblemBase, n_parallel: int = None, min_pof: float = None,
+                       g_aggregation: 'ConstraintAggregation' = None) -> Tuple['ConstrainedInfill', int]:
     """
     Get the default infill criterion according to the following logic:
     - If evaluations can be run in parallel:
@@ -55,7 +57,7 @@ def get_default_infill(problem: ArchOptProblemBase, n_parallel: int = None, min_
       - Multi-objective:  Ensemble of MPoI, MEPoI  with n_batch = 1
     - Set Probability of Feasibility as constraint handling technique if min_pof != .5, otherwise use g-mean prediction
 
-    Returns the infill, the recommended infill batch size, and whether constraints should be aggregated.
+    Returns the infill and recommended infill batch size.
     """
 
     # Determine number of evaluations that can be run in parallel
@@ -87,14 +89,12 @@ def get_default_infill(problem: ArchOptProblemBase, n_parallel: int = None, min_
     # Get infill and set constraint handling technique
     infill, n_batch = _get_infill()
 
-    aggregate_constraints = False
     if min_pof is not None and min_pof != .5:
-        aggregate_constraints = True
-        infill.constraint_strategy = ProbabilityOfFeasibility(min_pof=min_pof)
+        infill.constraint_strategy = ProbabilityOfFeasibility(min_pof=min_pof, aggregation=g_aggregation)
     else:
-        infill.constraint_strategy = MeanConstraintPrediction()
+        infill.constraint_strategy = MeanConstraintPrediction(aggregation=g_aggregation)
 
-    return infill, n_batch, aggregate_constraints
+    return infill, n_batch
 
 
 class SurrogateInfill:
@@ -128,6 +128,9 @@ class SurrogateInfill:
     @property
     def needs_variance(self):
         return False
+
+    def get_g_training_set(self, g: np.ndarray) -> np.ndarray:
+        return g
 
     def set_samples(self, x_train: np.ndarray, is_active_train: np.ndarray, y_train: np.ndarray):
         self.x_train = x_train
@@ -234,15 +237,54 @@ class FunctionEstimateInfill(SurrogateInfill):
[…]
