### 2023-06-21 75ebd88
Fix setup script
