+++ sb_arch_opt/tests/algo/test_pymoo.py
@@ -103,7 +103,7 @@ def test_store_results_restart():
             assert initialize_from_previous_results(nsga2, problem, tmp_folder) == (i > 0)
             if i > 0:
                 assert isinstance(nsga2.initialization.sampling, Population)
-                assert len(nsga2.initialization.sampling) == 100+2*100*i
+                assert len(nsga2.initialization.sampling) - (100+2*100*i) < 20
 
             minimize(problem, nsga2, termination=('n_gen', 3), copy_algorithm=False)
             assert os.path.exists(os.path.join(tmp_folder, 'pymoo_results.pkl'))
