### 2023-06-25 #5564 @athulpg007
Hi @jbussemaker, I tried to step through the [tutorial notebook](https://github.com/jbussemaker/SBArchOpt/blob/main/docs/tutorial.ipynb) here and ran into two issues.

**1. Issue with use of deprecated np.float**

```
from sb_arch_opt.problems.hierarchical import HierZDT1Small
problem = HierZDT1Small()
problem.print_stats()
```

In the print_stats line, an error occurs due to use [deprecated numpy.float](https://numpy.org/devdocs/release/1.20.0-notes.html#deprecations). Since numpy version is not pinned, I had the latest version 1.25.0. I suggest to replace the numpy.float with built-in float (preferable); or pin the numpy version to one < 1.20.0 (not recommended) as a short term fix. Full traceback:
```
Cell In[1], line 5
      3 # Instantiate the problem class and print some statistics
      4 problem = HierZDT1Small()
----> 5 problem.print_stats()

File ~/Documents/sbarchopt-test-py39/venv/lib/python3.9/site-packages/sb_arch_opt/problem.py:221, in ArchOptProblemBase.print_stats(self)
    218 print(f'MD     : {n_discr > 0 and n_cont > 0}')  # Is it a mixed-discrete problem?
    219 print(f'MO     : {self.n_obj > 1}')  # Is it a multi-objective problem?
--> 221 imp_ratio = self.get_imputation_ratio()
    222 discrete_imp_ratio = self.get_discrete_imputation_ratio()
    223 cont_imp_ratio = self.get_continuous_imputation_ratio()

File ~/Documents/sbarchopt-test-py39/venv/lib/python3.9/site-packages/sb_arch_opt/problem.py:244, in ArchOptProblemBase.get_imputation_ratio(self)
    239 def get_imputation_ratio(self) -> float:
    240     """
    241     Returns the problem-level imputation ratio, a measure of how hierarchical the problem is. It is calculated
    242     from the product of the discrete and continuous imputation ratios.
    243     """
--> 244     return self.design_space.imputation_ratio

File ~/Documents/sbarchopt-test-py39/venv/lib/python3.9/site-packages/cached_property.py:36, in cached_property.__get__(self, obj, cls)
     33 if asyncio and asyncio.iscoroutinefunction(self.func):
     34     return self._wrap_in_coroutine(obj)
---> 36 value = obj.__dict__[self.func.__name__] = self.func(obj)
     37 return value

File ~/Documents/sbarchopt-test-py39/venv/lib/python3.9/site-packages/sb_arch_opt/design_space.py:260, in ArchDesignSpace.imputation_ratio(self)
    254 @cached_property
    255 def imputation_ratio(self) -> float:
    256     """
    257     Returns the problem-level imputation ratio, a measure of how hierarchical the problem is. It is calculated
    258     from the product of the discrete and continuous imputation ratios.
    259     """
--> 260     return self.discrete_imputation_ratio * self.continuous_imputation_ratio

File ~/Documents/sbarchopt-test-py39/venv/lib/python3.9/site-packages/cached_property.py:36, in cached_property.__get__(self, obj, cls)
     33 if asyncio and asyncio.iscoroutinefunction(self.func):
     34     return self._wrap_in_coroutine(obj)
---> 36 value = obj.__dict__[self.func.__name__] = self.func(obj)
     37 return value

File ~/Documents/sbarchopt-test-py39/venv/lib/python3.9/site-packages/sb_arch_opt/design_space.py:278, in ArchDesignSpace.discrete_imputation_ratio(self)
    275 if n_valid == 0:
    276     return 1.
--> 278 n_declared = self.get_n_declared_discrete()
    279 discrete_imp_ratio = n_declared/n_valid
    280 return discrete_imp_ratio

File ~/Documents/sbarchopt-test-py39/venv/lib/python3.9/site-packages/sb_arch_opt/design_space.py:252, in ArchDesignSpace.get_n_declared_discrete(self)
    249 if len(n_opts_discrete) == 0:
    250     return 1
--> 252 return int(np.prod(n_opts_discrete, dtype=np.float))

File ~/Documents/sbarchopt-test-py39/venv/lib/python3.9/site-packages/numpy/__init__.py:313, in __getattr__(attr)
    308     warnings.warn(
    309         f"In the future `np.{attr}` will be defined as the "
    310         "corresponding NumPy scalar.", FutureWarning, stacklevel=2)
    312 if attr in __former_attrs__:
--> 313     raise AttributeError(__former_attrs__[attr])
    315 if attr == 'testing':
    316     import numpy.testing as testing

AttributeError: module 'numpy' has no attribute 'float'.
`np.float` was a deprecated alias for the builtin `float`. To avoid this error in existing code, use `float` by itself. Doing this will not modify any behavior and is safe. If you specifically wanted the numpy scalar type, use `np.float64` here.
The aliases was originally deprecated in NumPy 1.20; for more details and guidance see the original release note at:
    https://numpy.org/devdocs/release/1.20.0-notes.html#deprecations
```
**2. ValueError from pymoo (cause not known)**
```
import tempfile

# We create a temporary folder for demonstration purposes
td = tempfile.TemporaryDirectory()
results_folder = td.name
# Create the algorithm with the results folder
nsga2 = get_nsga2(pop_size=100, results_folder=results_folder)
# Run for a few generations
result5 = minimize(problem, nsga2, termination=('n_gen', 5), copy_algorithm=False)
```
```
Cell In[5], line 11
      8 nsga2 = get_nsga2(pop_size=100, results_folder=results_folder)
     10 # Run for a few generations
---> 11 result5 = minimize(problem, nsga2, termination=('n_gen', 5), copy_algorithm=False)
     13 assert nsga2.evaluator.n_eval == 500  # 5*100
     14 plot(problem.pareto_front(), result5.opt.get('F'),
     15      labels=['Known Pareto front', 'NSGA2 (5 gen)'])

File ~/Documents/sbarchopt-test-py39/venv/lib/python3.9/site-packages/pymoo/optimize.py:67, in minimize(problem, algorithm, termination, copy_algorithm, copy_termination, **kwargs)
     64     algorithm.setup(problem, **kwargs)
     66 # actually execute the algorithm
---> 67 res = algorithm.run()
     69 # store the deep copied algorithm in the result object
     70 res.algorithm = algorithm

File ~/Documents/sbarchopt-test-py39/venv/lib/python3.9/site-packages/pymoo/core/algorithm.py:141, in Algorithm.run(self)
    139 def run(self):
    140     while self.has_next():
--> 141         self.next()
    142     return self.result()

File ~/Documents/sbarchopt-test-py39/venv/lib/python3.9/site-packages/pymoo/core/algorithm.py:161, in Algorithm.next(self)
    159 # call the advance with them after evaluation
    160 if infills is not None:
--> 161     self.evaluator.eval(self.problem, infills, algorithm=self)
    162     self.advance(infills=infills)
    164 # if the algorithm does not follow the infill-advance scheme just call advance
    165 else:

File ~/Documents/sbarchopt-test-py39/venv/lib/python3.9/site-packages/sb_arch_opt/algo/pymoo_interface/storage_restart.py:208, in ArchOptEvaluator.eval(self, problem, pop, skip_already_evaluated, evaluate_values_of, count_evals, **kwargs)
    205     is_duplicate = LargeDuplicateElimination.eliminate(self._cumulative_pop.get('X'), pop.get('X'))
    206     self._non_eval_cumulative_pop = self._cumulative_pop[~is_duplicate]
--> 208 results = super().eval(problem, pop, skip_already_evaluated=skip_already_evaluated,
    209                        evaluate_values_of=evaluate_values_of, count_evals=count_evals, **kwargs)
    211 # Post-evaluation storage
    212 if self.results_folder is not None:

File ~/Documents/sbarchopt-test-py39/venv/lib/python3.9/site-packages/pymoo/core/evaluator.py:69, in Evaluator.eval(self, problem, pop, skip_already_evaluated, evaluate_values_of, count_evals, **kwargs)
     65 # evaluate the solutions (if there are any)
     66 if len(I) > 0:
     67
     68     # do the actual evaluation - call the sub-function to set the corresponding values to the population
---> 69     self._eval(problem, pop[I], evaluate_values_of, **kwargs)
     71 # update the function evaluation counter
     72 if count_evals:

File ~/Documents/sbarchopt-test-py39/venv/lib/python3.9/site-packages/sb_arch_opt/algo/pymoo_interface/storage_restart.py:239, in ArchOptEvaluator._eval(self, problem, pop, evaluate_values_of, **kwargs)
    236         super()._eval(problem, batch_pop, evaluate_values_of, **kwargs)
    238         self._apply_extreme_barrier(batch_pop)
--> 239         intermediate_pop = self._normalize_pop(pop, evaluate_values_of, evaluated_pop=self._evaluated_pop)
    240         self._store_intermediate(problem, intermediate_pop)
    242 # Apply extreme barrier: replace NaN with Inf

File ~/Documents/sbarchopt-test-py39/venv/lib/python3.9/site-packages/sb_arch_opt/algo/pymoo_interface/storage_restart.py:259, in ArchOptEvaluator._normalize_pop(pop, evaluate_values_of, nan_as_inf, evaluated_pop)
    257 pop_data = {}
    258 for key in (['X']+evaluate_values_of):
--> 259     data = pop.get(key)
    261     if len(data.shape) == 1:
    262         partial_data = np.zeros((len(data), len(data[0])))*np.nan

File ~/Documents/sbarchopt-test-py39/venv/lib/python3.9/site-packages/pymoo/core/population.py:67, in Population.get(self, to_numpy, *args, **kwargs)
     65 # to numpy array if desired - default true
     66 if to_numpy:
---> 67     res = [np.array(e) for e in res]
     69 # return as tuple or single value
     70 if len(args) == 1:

File ~/Documents/sbarchopt-test-py39/venv/lib/python3.9/site-packages/pymoo/core/population.py:67, in <listcomp>(.0)
     65 # to numpy array if desired - default true
     66 if to_numpy:
---> 67     res = [np.array(e) for e in res]
     69 # return as tuple or single value
     70 if len(args) == 1:

ValueError: setting an array element with a sequence. The requested array has an inhomogeneous shape after 1 dimensions. The detected shape was (100,) + inhomogeneous part.
```
