### 2023-06-25 b5b5adc
Replace deprecated np.float
