### 2023-06-25 dd43a7c
Fix intermediate storage
