### 2023-06-25 8ccfd8d
Add numba to dependencies
