+++ .github/workflows/tests.yml
@@ -35,7 +35,8 @@ jobs:
 
     - name: Install dependencies
       run: |
-        python -m pip install --upgrade pip setuptools pytest
+        python -m pip install --upgrade pip setuptools
+        pip install -r requirements-tests.txt
         pip install -r requirements-assignment.txt
         pip install -r requirements-hebo.txt
         pip install -e .[arch_sbo,botorch,trieste,tpe]
