### 2023-06-26 0f941fc
Use Numba for SMT Kriging models
