+++ .github/workflows/tests_slow.yml
@@ -1,7 +1,7 @@
 # This workflow will install Python dependencies, run tests and lint with a variety of Python versions
 # For more information see: https://docs.github.com/en/actions/automating-builds-and-tests/building-and-testing-python
 
-name: Tests
+name: Slow Tests
 
 on:
   push:
@@ -15,11 +15,8 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        os: [ubuntu-latest, macos-latest, windows-latest]
-        python-version: ["3.10"]
-        include:
-          - os: ubuntu-latest
-            python-version: "3.8"
+        os: [ubuntu-latest]
+        python-version: ["3.9"]
 
     steps:
     - uses: actions/checkout@v3
@@ -40,4 +37,4 @@ jobs:
         pip install -e .[arch_sbo,botorch,trieste,tpe]
 
     - name: Test with pytest
-      run: pytest sb_arch_opt --durations=0
+      run: RUN_SLOW_TESTS=1 pytest sb_arch_opt --durations=0
