### 2023-06-26 421059e
Fix partial storage
