### 2023-06-26 #5564 @jbussemaker
@athulpg007 thank you for catching these problems. I have corrected them, in addition to adding unit tests to also test the tutorial notebooks.

I also corrected another small mistake I made with one of the dependencies I use: if you're starting from a previously-prepared Python environment, ensure you have numba installed: `pip install numba`

Thank you for your time reviewing the library!
