### 2023-07-03 6cf2b79
Interface for passing kwargs to SMT models
