### 2023-07-04 jbussemaker/SBArchOpt#4 @athulpg007
Failure in assert HAS_ARCH_SBO
When running the following line in the tutorial notebook, I encounter an assertion error with no message on the failure.
`from sb_arch_opt.algo.arch_sbo import get_arch_sbo_gp`
```
AssertionError                            Traceback (most recent call last)
Cell In[15], line 1
----> 1 from sb_arch_opt.algo.arch_sbo import get_arch_sbo_gp
      3 n_doe = 10
      4 n_infill = 20

File ~/Documents/sbarchopt/venv/lib/python3.10/site-packages/sb_arch_opt/algo/arch_sbo/__init__.py:1
----> 1 from .api import *

File ~/Documents/sbarchopt/venv/lib/python3.10/site-packages/sb_arch_opt/algo/arch_sbo/api.py:27
     25 from sb_arch_opt.problem import *
     26 from sb_arch_opt.algo.arch_sbo.models import *
---> 27 from sb_arch_opt.algo.arch_sbo.algo import *
     28 from sb_arch_opt.algo.arch_sbo.infill import *
     29 from sb_arch_opt.algo.arch_sbo.metrics import *

File ~/Documents/sbarchopt/venv/lib/python3.10/site-packages/sb_arch_opt/algo/arch_sbo/algo.py:55
     53 try:
     54     from smt.surrogate_models.surrogate_model import SurrogateModel
---> 55     from sb_arch_opt.algo.arch_sbo.infill import *
     56     from sb_arch_opt.algo.arch_sbo.models import *
     57     from sb_arch_opt.algo.arch_sbo.hc_strategy import *

File ~/Documents/sbarchopt/venv/lib/python3.10/site-packages/sb_arch_opt/algo/arch_sbo/infill.py:32
     30 from scipy.special import ndtr
     31 from sb_arch_opt.problem import ArchOptProblemBase
---> 32 from sb_arch_opt.algo.arch_sbo.hc_strategy import HiddenConstraintStrategy
     34 from pymoo.core.problem import Problem
     35 from pymoo.core.population import Population

File ~/Documents/sbarchopt/venv/lib/python3.10/site-packages/sb_arch_opt/algo/arch_sbo/hc_strategy.py:33
     31 try:
     32     from smt.surrogate_models.surrogate_model import SurrogateModel
---> 33     assert HAS_ARCH_SBO
     34 except ImportError:
     35     assert not HAS_ARCH_SBO

AssertionError:
```
