### 2023-07-04 jbussemaker/SBArchOpt#5 @athulpg007
Consider adding API reference?
Would it help users to have an API reference for SBArchOpt such as these?
1. https://pymoo.org/api/algorithms.html
2. https://secondmind-labs.github.io/trieste/1.1.2/autoapi/trieste/bayesian_optimizer/index.html
