### 2023-07-04 jbussemaker/SBArchOpt#4 @athulpg007
Occurs because HAS_ARCH_SBO is None?
