### 2023-07-04 jbussemaker/SBArchOpt#6 @athulpg007
Consider adding a minimal example for a new user to get started
The currently available tutorial notebooks are great. However, I think it would help to have a minimal smaller example (if possible) to help a new user get familiar with how SBArchOpt works, and potentially highlight its capabilities over other tools.
