### 2023-07-04 #5564 @athulpg007
@jbussemaker please take a look: https://github.com/jbussemaker/SBArchOpt/issues/4, https://github.com/jbussemaker/SBArchOpt/issues/5, https://github.com/jbussemaker/SBArchOpt/issues/6.
