### 2023-07-04 jbussemaker/SBArchOpt#4 @athulpg007
This was resolved, I am not sure why I got the error earlier.
