### 2023-07-04 jbussemaker/SBArchOpt#6 @athulpg007
Alternatively, consider adding some more explanation(s) of what each step of the existing tutorial notebook is doing to walk the user through the steps.
