### 2023-07-05 eef0387
Upgrade smt version
