+++ setup.py
@@ -62,7 +62,7 @@ if __name__ == '__main__':
         ],
         extras_require={
             'arch_sbo': [
-                'smt~=2.0b5',
+                'smt~=2.0',
                 'numba',
                 'scikit-learn',
             ],
