### 2023-07-05 7ce0b68
Fixed KPLS usage for mixed-discrete/hierarchical problems
