### 2023-07-05 d6e7311
Increase tests speed
