### 2023-07-05 7670415
Add standalone minimal examples (#6)
