+++ docs/api/trieste.md
@@ -1,0 +1,14 @@
+# Trieste API Reference
+
+[Installation and usage](../algo/trieste.md)
+
+::: sb_arch_opt.algo.trieste_interface.api.get_trieste_optimizer
+    handler: python
+
+::: sb_arch_opt.algo.trieste_interface.algo.ArchOptBayesianOptimizer
+    handler: python
+    options:
+        members:
+            - initialize_from_previous
+            - run_optimization
+            - to_population

[…]
