### 2023-07-05 jbussemaker/SBArchOpt#5 @jbussemaker
@athulpg007 thank you for the suggestion, I have upgraded the docs to a static website (including API reference): https://sbarchopt.readthedocs.io/
