### 2023-07-08 #5564 @athulpg007
Hi @jbussemaker, thank you for addressing my comments. I have completed my review and am happy to accept this paper.

A few minor suggestions:

1. Please update the [PyPI](https://pypi.org/project/sb-arch-opt/) release to the latest version which includes the recent bug fixes.
2. 'An accompanying publication will be prepared in the near future detailing more about the implemented optimization features.' I suggest to remove this line from the paper, as it is making a promise of future work that is not yet done at this time. Once the mentioned publication is available, it can be linked to in the documentation page. This will allow users who want to gain a more in depth understanding of the features to refer to it. You can also consider linking any other published articles which used the library.

Thank you for this valuable contribution to the field of systems architecture optimization.
