### 2023-07-08 bcdd3e0
Remove future paper line
