### 2023-07-19 cc18c81
Add test without optional dependencies
