### 2023-07-24 97f1a26
Use exact checks with seed
