+++ sb_arch_opt/tests/algo/test_arch_sbo.py
@@ -337,8 +337,13 @@ def test_smt_krg_features():
         categorical_kernel=MixIntKernelType.EXP_HOMO_HSPHERE,
         hierarchical_kernel=MixHrcKernelType.ALG_KERNEL,
     )
+    pls_kwargs = dict(
+        categorical_kernel=MixIntKernelType.CONT_RELAX,
+        hierarchical_kernel=MixHrcKernelType.ALG_KERNEL,
+    )
 
-    def _try_model(problem: ArchOptProblemBase, pls: bool = False, cont_relax: bool = False, throws_error=False):
+    def _try_model(problem: ArchOptProblemBase, pls: bool = False, cont_relax: bool = False, ignore_hierarchy=False,
+                   throws_error=False):
         model_factory = ModelFactory(problem)
         normalization = model_factory.get_md_normalization()
         ds = model_factory.problem.design_space
@@ -347,15 +352,17 @@ def test_smt_krg_features():
         if cont_relax and norm_ds_spec.is_mixed_discrete:
             cr_ds_spec = model_factory.create_smt_design_space_spec(ds, md_normalize=True, cont_relax=True)
             assert cr_ds_spec.design_space.is_all_cont
-            if pls:
-                model = KPLS(n_comp=n_pls, design_space=cr_ds_spec.design_space, **kwargs)
-            else:
-                model = KRG(design_space=cr_ds_spec.design_space, **kwargs)
+            model_ds = cr_ds_spec.design_space
+        elif ignore_hierarchy:
+            model_ds = model_factory.create_smt_design_space_spec(
+                ds, md_normalize=True, ignore_hierarchy=True).design_space
         else:
-            if pls:
-                model = KPLS(n_comp=n_pls, design_space=norm_ds_spec.design_space, **kwargs)
-            else:
-                model = KRG(design_space=norm_ds_spec.design_space, **kwargs)
+            model_ds = norm_ds_spec.design_space
+
+        if pls:
+            model = KPLS(n_comp=n_pls, design_space=model_ds, **pls_kwargs)
+        else:
+            model = KRG(design_space=model_ds, **kwargs)
 
         model = MultiSurrogateModel(model)
 
@@ -375,34 +382,40 @@ def test_smt_krg_features():
         _try_model(Rosenbrock(), cont_relax=True)
         _try_model(Rosenbrock())
         _try_model(Rosenbrock(), pls=True, cont_relax=True)
+        _try_model(Rosenbrock(), pls=True, ignore_hierarchy=True)
         _try_model(Rosenbrock(), pls=True)
 
         # Mixed-discrete (integer)
         _try_model(MDMORosenbrock(), cont_relax=True)
         _try_model(MDMORosenbrock())
         _try_model(MDMORosenbrock(), pls=True, cont_relax=True)
+        _try_model(MDMORosenbrock(), pls=True, ignore_hierarchy=True)
         _try_model(MDMORosenbrock(), pls=True)
 
         # Mixed-discrete (categorical)
         _try_model(Halstrup04(), cont_relax=True)
         _try_model(Halstrup04())
         _try_model(Halstrup04(), pls=True, cont_relax=True)
+        _try_model(Halstrup04(), pls=True, ignore_hierarchy=True)
         _try_model(Halstrup04(), pls=True)
 
         # Hierarchical (continuous conditional vars)
         _try_model(ZaeffererHierarchical(), cont_relax=True)
         _try_model(ZaeffererHierarchical())
         _try_model(ZaeffererHierarchical(), pls=True, cont_relax=True)
+        _try_model(ZaeffererHierarchical(), pls=True, ignore_hierarchy=True)
[…]
