### 2023-08-11 0e6a18d
Support dynamic nr of trees in random forest classifier
