### 2023-08-11 59e1b30
Fix jupyter kernel install for tests
