+++ .github/workflows/tests_slow.yml
@@ -35,6 +35,8 @@ jobs:
         pip install -r requirements-assignment.txt
         pip install -r requirements-hebo.txt
         pip install -e .[arch_sbo,botorch,trieste,tpe]
+        pip install jupyter ipython ipykernel
+        ipython kernel install --name "python3" --user
 
     - name: Test with pytest
       run: RUN_SLOW_TESTS=1 pytest sb_arch_opt --durations=0
