### 2023-09-01 140dde2
Support categorical decreed vars in SMT 2.1
