### 2023-09-01 ef0b842
Support changing DOE size
