### 2023-09-01 c46cf05
Expose storage restart functionality for ask-tell
