+++ sb_arch_opt/algo/pymoo_interface/storage_restart.py
@@ -46,17 +46,20 @@ __all__ = ['load_from_previous_results', 'initialize_from_previous_results', 'Re
 log = logging.getLogger('sb_arch_opt.pymoo')
 
 
-def load_from_previous_results(problem: ArchOptProblemBase, result_folder: str) -> Optional[Population]:
+def load_from_previous_results(problem: ArchOptProblemBase, result_folder: str, load_from_problem=True,
+                               cumulative=True) -> Optional[Population]:
     """Load a (cumulative) Population from previously-stored results"""
     capture_log()
 
     # Try to load using problem-specific function first
-    population = problem.load_previous_results(result_folder)
-    if population is not None and len(population) > 0:
-        log.info(f'Previous results loaded from problem results: {len(population)} design points')
+    population = None
+    if load_from_problem:
+        population = problem.load_previous_results(result_folder)
+        if population is not None and len(population) > 0:
+            log.info(f'Previous results loaded from problem results: {len(population)} design points')
 
     # Additionally try to load from pymoo storage to merge with non-evaluated design points
-    pymoo_population = ArchOptEvaluator.load_pop(result_folder)
+    pymoo_population = ArchOptEvaluator.load_pop(result_folder, cumulative=cumulative)
     if pymoo_population is not None and len(pymoo_population) > 0:
 
         if population is None:
@@ -89,7 +92,8 @@ def load_from_previous_results(problem: ArchOptProblemBase, result_folder: str) 
     return population
 
 
-def initialize_from_previous_results(algorithm: Algorithm, problem: ArchOptProblemBase, result_folder: str) -> bool:
+def initialize_from_previous_results(algorithm: Algorithm, problem: ArchOptProblemBase, result_folder: str, **kwargs) \
+        -> bool:
     """Initialize an Algorithm from previously stored results"""
     capture_log()
 
@@ -97,10 +101,19 @@ def initialize_from_previous_results(algorithm: Algorithm, problem: ArchOptProbl
         raise RuntimeError(f'Algorithm has no initialization step, cannot set initial population: {algorithm!r}')
 
     # Try to load from previous results
-    population = load_from_previous_results(problem, result_folder)
+    population = load_from_previous_results(problem, result_folder, **kwargs)
     if population is None:
         log.info(f'No previous population found, not changing initialization strategy')
         return False
+
+    # Set static initialization on the algorithm to start from the loaded population
+    set_initial_population(algorithm, population)
+
+    return True
+
+
+def set_initial_population(algorithm: Algorithm, population: Population):
+    """Set the initial population on an Algorithm"""
 
     # Set static initialization on the algorithm to start from the loaded population
     algorithm.initialization = Initialization(population)
@@ -108,8 +121,6 @@ def initialize_from_previous_results(algorithm: Algorithm, problem: ArchOptProbl
     # Initialize cumulative population
     if isinstance(algorithm.evaluator, ArchOptEvaluator):
         algorithm.evaluator.initialize_cumulative(population)
-
-    return True
 
 
 class ResultsStorageCallback(Callback):
@@ -169,6 +180,22 @@ class ArchOptEvaluator(Evaluator):
     Batch process size is determined using `get_n_batch_evaluate` if not specified explicitly!
 
     Also using the ResultsStorageCallback ensures that also final problem-specific results are stored.
+
+    You can also use this class with intermediate (partial) results storage in an ask-tell setting:
+    >>> algo: Algorithm
+    >>> problem: ArchOptProblemBase
[…]
