### 2023-09-05 #5564 @jbussemaker
@vissarion thank you.
The release is v1.2.0 and is archived at [10.5281/zenodo.8318765](https://doi.org/10.5281/zenodo.8318765)
