+++ paper.bib
@@ -63,6 +63,7 @@ year = {2015}
   year      = {2023},
   address   = {Honolulu, HI, USA},
   month     = jul,
+  doi       = {10.1002/iis2.13020},
 }
 
 @InProceedings{Bussemaker2022,
@@ -72,6 +73,7 @@ year = {2015}
   year      = {2022},
   address   = {Detroit, MI, USA},
   month     = jun,
+  doi       = {10.1002/iis2.12935},
 }
 
 @inproceedings{Balandat2020,
