### 2023-09-05 #5564 @jbussemaker
> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> OK DOIs
>
> - 10.1007/978-1-4020-4399-4 is OK
> - 10.1007/978-3-030-27486-3_36-1 is OK
> - 10.2514/6.2021-3095 is OK
> - 10.1007/s00500-017-2965-0 is OK
> - 10.1017/9781108348973 is OK
> - 10.48550/ARXIV.2302.08436 is OK
> - 10.1109/access.2020.2990567 is OK
> - 10.1016/j.ast.2019.03.041 is OK
> - 10.2514/6.2022-3899 is OK
> - 10.1115/detc2021-71399 is OK
> - 10.2514/6.2021-3078 is OK
>
> MISSING DOIs
>
> - 10.1002/iis2.13020 may be a valid DOI for title: Function-Based Architecture Optimization: An Application to Hybrid-Electric Propulsion Systems
> - 10.1002/iis2.12935 may be a valid DOI for title: From System Architecting to System Design and Optimization: A Link Between MBSE and MDAO
>
> INVALID DOIs
>
> - None
> ```

I have added the DOIs. Somehow they didn't work previously...
