### 2023-09-06 jbussemaker/SBArchOpt#7 @vissarion
JOSS references
Two comments on JOSS references. Related to JOSS review https://github.com/openjournals/joss-reviews/issues/5564

For reference (Cowen-Rivers et al. 2019) this DOI https://doi.org/10.1613/jair.1.13643 can be used.

Please use "{}" to force capitalization in references e.g. edit "monte carlo" to "Monte Carlo"
