### 2023-09-06 249bc99
Add DOI and fix capitalization (#7)
