+++ paper.bib
@@ -13,7 +13,7 @@ year = {2015}
   author    = {J.H. Bussemaker and P.D. Ciampa},
   booktitle = {Handbook of Model-Based Systems Engineering},
   publisher = {Springer},
-  title     = {MBSE in Architecture Design Space Exploration},
+  title     = {{MBSE in Architecture Design Space Exploration}},
   year      = {2022},
   editor    = {A.M. Madni and N. Augustine and M. Sievers},
   month     = {April},
@@ -25,7 +25,7 @@ year = {2015}
 @InProceedings{Bussemaker2021,
   author    = {J.H. Bussemaker and N. Bartoli and T. Lefebvre and P.D. Ciampa and B. Nagel},
   booktitle = {American Institute of Aeronautics and Astronautics ({AIAA}) {AVIATION} 2021 {FORUM}},
-  title     = {Effectiveness of Surrogate-Based Optimization Algorithms for System Architecture Optimization},
+  title     = {{Effectiveness of Surrogate-Based Optimization Algorithms for System Architecture Optimization}},
   year      = {2021},
   address   = {Virtual Event},
   month     = aug,
@@ -35,7 +35,7 @@ year = {2015}
 @Article{Chugh2019,
   author    = {Chugh, T. and Sindhya, K. and Hakanen, J. and Miettinen, K.},
   journal   = {Soft Computing},
-  title     = {A survey on handling computationally expensive multiobjective optimization problems with evolutionary algorithms},
+  title     = {{A survey on handling computationally expensive multiobjective optimization problems with evolutionary algorithms}},
   year      = {2019},
   issn      = {1432-7643},
   month     = {may},
@@ -50,7 +50,7 @@ year = {2015}
 @Book{Garnett2023,
   author    = {Garnett, R.},
   publisher = {Cambridge University Press},
-  title     = {Bayesian Optimization},
+  title     = {{Bayesian Optimization}},
   year      = {2023},
   month     = {jan},
   doi       = {10.1017/9781108348973},
@@ -59,7 +59,7 @@ year = {2015}
 @InProceedings{Bussemaker2023,
   author    = {Bussemaker, J.H. and García Sánchez, R. and Fouda, M. and Boggero, L. and Nagel, B.},
   booktitle = {33rd Annual International Council on Systems Engineering ({INCOSE}) International Symposium},
-  title     = {Function-Based Architecture Optimization: An Application to Hybrid-Electric Propulsion Systems},
+  title     = {{Function-Based Architecture Optimization: An Application to Hybrid-Electric Propulsion Systems}},
   year      = {2023},
   address   = {Honolulu, HI, USA},
   month     = jul,
@@ -69,7 +69,7 @@ year = {2015}
 @InProceedings{Bussemaker2022,
   author    = {Bussemaker, J.H. and Boggero, L. and Ciampa, P. D.},
   booktitle = {32nd Annual International Council on Systems Engineering ({INCOSE}) International Symposium},
-  title     = {From System Architecting to System Design and Optimization: A Link Between {MBSE} and {MDAO}},
+  title     = {{From System Architecting to System Design and Optimization: A Link Between MBSE and MDAO}},
   year      = {2022},
   address   = {Detroit, MI, USA},
   month     = jun,
@@ -77,7 +77,7 @@ year = {2015}
 }
 
 @inproceedings{Balandat2020,
-  title = {BoTorch: A Framework for Efficient Monte-Carlo Bayesian Optimization},
[…]
