+++ paper.bib
@@ -97,7 +97,7 @@ year = {2015}
 }
 
 @article{CowenRivers2022,
-  title={{HEBO: pushing the limits of sample-efficient hyper-parameter optimisation},
+  title={{HEBO: pushing the limits of sample-efficient hyper-parameter optimisation}},
   author={Cowen-Rivers, Alexander I and Lyu, Wenlong and Tutunov, Rasul and Wang, Zhi and Grosnit, Antoine and Griffiths, Ryan Rhys and Maraval, Alexandre Max and Jianye, Hao and Wang, Jun and Peters, Jan and others},
   journal={Journal of Artificial Intelligence Research},
   volume={74},
