### 2023-09-07 #5564 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
